login-service/pyproject.toml
[project]
name = "login-service"
version = "0.1.0"
description = "Login Service for sia-helpers"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "mouseinfo>=0.1.3",
    "pyautogui>=0.9.54",
    "pygetwindow>=0.0.9",
    "pymsgbox>=1.0.9",
    "pyperclip>=1.9.0",
    "pyrect>=0.2.0",
    "pyscreeze>=1.0.1",
    "python-dotenv>=1.1.1",
    "python3-xlib>=0.15",
    "pytweening>=1.2.0",
    "seleniumbase>=4.39.5",
    "uvicorn[standard]>=0.34.3",
]
